requirements.txt
PythonToolsKit
pycryptodome